work_board/table/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-taggit==6.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
sqlparse==0.5.1
tzdata==2024.2